+++ src/Microplate/Coordinates.php
@@ -50,11 +50,11 @@ class Coordinates
      * @param CoordinatesArray $coordinates
      * @param TCoord $coordinateSystem
      *
-     * @return self<TCoord>
+     * @return static<TCoord>
      */
-    public static function fromArray(array $coordinates, CoordinateSystem $coordinateSystem): self
+    public static function fromArray(array $coordinates, CoordinateSystem $coordinateSystem): static
     {
-        return new self($coordinates['row'], $coordinates['column'], $coordinateSystem);
+        return new static($coordinates['row'], $coordinates['column'], $coordinateSystem);
     }
 
     /**
@@ -62,9 +62,9 @@ class Coordinates
      *
      * @param TCoord $coordinateSystem
      *
-     * @return self<TCoord>
+     * @return static<TCoord>
      */
-    public static function fromString(string $coordinatesString, CoordinateSystem $coordinateSystem): self
+    public static function fromString(string $coordinatesString, CoordinateSystem $coordinateSystem): static
     {
         $rows = $coordinateSystem->rows();
         $rowsOptions = \implode('|', $rows);
@@ -88,7 +88,7 @@ class Coordinates
             throw new \InvalidArgumentException("Expected coordinates between {$firstValidExample} and {$lastValidExample} for {$coordinateSystemClass}, got: {$coordinatesString}.");
         }
 
-        return new self($matches[1], (int) $matches[2], $coordinateSystem);
+        return new static($matches[1], (int) $matches[2], $coordinateSystem);
     }
 
     /**
@@ -96,22 +96,22 @@ class Coordinates
      *
      * @param TCoord $coordinateSystem
      *
-     * @return self<TCoord>
+     * @return static<TCoord>
      */
-    public static function fromPosition(int $position, FlowDirection $direction, CoordinateSystem $coordinateSystem): self
+    public static function fromPosition(int $position, FlowDirection $direction, CoordinateSystem $coordinateSystem): static
     {
         self::assertPositionInRange($coordinateSystem, $position);
 
         switch ($direction->value) {
             case FlowDirection::COLUMN:
-                return new self(
+                return new static(
                     $coordinateSystem->rowForColumnFlowPosition($position),
                     $coordinateSystem->columnForColumnFlowPosition($position),
                     $coordinateSystem
                 );
 
             case FlowDirection::ROW:
-                return new self(
+                return new static(
                     $coordinateSystem->rowForRowFlowPosition($position),
                     $coordinateSystem->columnForRowFlowPosition($position),
                     $coordinateSystem