+++ src/Tecan/LiquidClass/MLLLiquidClass.php
@@ -17,29 +17,29 @@ class MLLLiquidClass implements LiquidClass
         $this->value = $value;
     }
 
-    public static function DNA_DILUTION(): self
+    public static function DNA_DILUTION(): static
     {
-        return new self(self::DNA_DILUTION);
+        return new static(self::DNA_DILUTION);
     }
 
-    public static function DNA_DILUTION_WATER(): self
+    public static function DNA_DILUTION_WATER(): static
     {
-        return new self(self::DNA_DILUTION_WATER);
+        return new static(self::DNA_DILUTION_WATER);
     }
 
-    public static function TRANSFER_PCR_PRODUKT(): self
+    public static function TRANSFER_PCR_PRODUKT(): static
     {
-        return new self(self::TRANSFER_PCR_PRODUKT);
+        return new static(self::TRANSFER_PCR_PRODUKT);
     }
 
-    public static function TRANSFER_MASTERMIX_MP(): self
+    public static function TRANSFER_MASTERMIX_MP(): static
     {
-        return new self(self::TRANSFER_MASTERMIX_MP);
+        return new static(self::TRANSFER_MASTERMIX_MP);
     }
 
-    public static function TRANSFER_TEMPLATE(): self
+    public static function TRANSFER_TEMPLATE(): static
     {
-        return new self(self::TRANSFER_TEMPLATE);
+        return new static(self::TRANSFER_TEMPLATE);
     }
 
     public function name(): string