+++ src/Tecan/TipMask/TipMask.php
@@ -18,14 +18,14 @@ class TipMask
         $this->value = $value;
     }
 
-    public static function FOUR_TIPS(): self
+    public static function FOUR_TIPS(): static
     {
-        return new self(self::FOUR_TIPS);
+        return new static(self::FOUR_TIPS);
     }
 
-    public static function EIGHT_TIPS(): self
+    public static function EIGHT_TIPS(): static
     {
-        return new self(self::EIGHT_TIPS);
+        return new static(self::EIGHT_TIPS);
     }
 
     public static function firstTip(): int